requirements.txt
streamlit==1.18.0
pyperclip==1.8.2
gitpython==3.1.31
